requirements.txt
django
djangorestframework
coverage